refactor: move tokenize and fix comment handling
